Re-enabled additional tests
